% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.flexmirt.R
\name{extract.flexmirt}
\alias{extract.flexmirt}
\title{Extract flexmirt estimation}
\usage{
\method{extract}{flexmirt}(filename)
}
\arguments{
\item{filename}{the path flexmirt output file}
}
\description{
Extract flexmirt estimation
}
