% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IRTmixed.R
\name{IRTmixed}
\alias{IRTmixed}
\title{main function to fit the IRT model using mixed model}
\usage{
IRTmixed(x, itemtype = "2PL", a.fix = 1)
}
\arguments{
\item{x}{a data.frame object. Each row is observarion, and each column is one item.}
}
\description{
main function to fit the IRT model using mixed model
}
