% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.irtmixed.R
\name{plot.irtmixed}
\alias{plot.irtmixed}
\title{main function to fit the IRT model using mixed model}
\usage{
plot.irtmixed(x, type = "irf", ...)
}
\arguments{
\item{x}{a irtmixed object. Each row is observarion, and each column is one item.}
}
\description{
main function to fit the IRT model using mixed model
}
