% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.irtmixed.R
\name{summary.irtmixed}
\alias{summary.irtmixed}
\title{summarize the coefficients of irtmixed object}
\usage{
\method{summary}{irtmixed}(x)
}
\arguments{
\item{x}{a irtmixed object. Each row is observarion, and each column is one item.}
}
\description{
summarize the coefficients of irtmixed object
}
