% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write.DTflexmirt.R
\name{write.DTflexmirt}
\alias{write.DTflexmirt}
\title{Reshape the data and save sas the format for flexmirt}
\usage{
write.DTflexmirt(x, file)
}
\arguments{
\item{x}{a data.frame object. Each row is observarion, and each column is one item.}

\item{file}{the path dataset save}
}
\description{
Reshape the data and save sas the format for flexmirt
}
